+++ types/protocol-mapping.d.ts
@@ -559,6 +559,11 @@ export namespace ProtocolMapping {
         'Storage.storageBucketCreatedOrUpdated': [Protocol.Storage.StorageBucketCreatedOrUpdatedEvent];
         'Storage.storageBucketDeleted': [Protocol.Storage.StorageBucketDeletedEvent];
         /**
+         * TODO(crbug.com/1458532): Add other Attribution Reporting events, e.g.
+         * trigger registration.
+         */
+        'Storage.attributionReportingSourceRegistered': [Protocol.Storage.AttributionReportingSourceRegisteredEvent];
+        /**
          * Issued when attached to target because of auto-attach or `attachToTarget` command.
          */
         'Target.attachedToTarget': [Protocol.Target.AttachedToTargetEvent];
@@ -4277,6 +4282,13 @@ export namespace ProtocolMapping {
          */
         'Storage.setAttributionReportingLocalTestingMode': {
             paramsType: [Protocol.Storage.SetAttributionReportingLocalTestingModeRequest];
+            returnType: void;
+        };
+        /**
+         * Enables/disables issuing of Attribution Reporting events.
+         */
+        'Storage.setAttributionReportingTracking': {
+            paramsType: [Protocol.Storage.SetAttributionReportingTrackingRequest];
             returnType: void;
         };
         /**